fix: dual-i2c offset sync and stack safety
- pendingCharOffsetsMask/charMask uint8_t -> uint16_t: bits 8-15
  truncated, so char-offset pushes for modules 8-15 (bus 2) were
  silently dropped on 16-module slaves
- applyOffsetsPush now splits the flat vector: 0-7 -> moduleOffsets,
  8-15 -> wire1Offsets (dual builds), matching reloadOffsets()'s
  per-bus reads; previously bus-2 offsets were stored where they
  were never read
- reportOffsetsToMaster merges both bus vectors and fills all
  moduleCount slots; previously it sent moduleCount=16 with only
  8 offsets filled (rest zero), so the master stored 0s for bus 2
- distributeMessage: String chunks[maxChunks] stack VLA (384 x ~16B
  = ~6KB on dual builds vs 8KB loop stack) -> std::vector heap alloc

Host tests extended with mirrors of the push-split and report-merge.

+++ src/SplitFlapEspNow.h
@@ -109,7 +109,7 @@ class SplitFlapEspNow {
 
     volatile bool pendingOffsetsPush;
     SplitFlapOffsetsPushMessage pendingOffsetsPushPkt;
-    volatile uint8_t pendingCharOffsetsMask;
+    volatile uint16_t pendingCharOffsetsMask;
     SplitFlapCharOffsetsPushMessage pendingCharOffsetsPkts[MAX_MODULES];
     bool offsetDataDirty;
     unsigned long lastOffsetRxMs;
